Moves autogenerated code for peripherals into seperate files & Changes measurement units to avoid using floats
