+++ weatherballoon.ioc
@@ -220,7 +220,7 @@ ProjectManager.AskForMigrate=true
 ProjectManager.BackupPrevious=false
 ProjectManager.CompilerOptimize=6
 ProjectManager.ComputerToolchain=false
-ProjectManager.CoupleFile=false
+ProjectManager.CoupleFile=true
 ProjectManager.CustomerFirmwarePackage=
 ProjectManager.DefaultFWLocation=true
 ProjectManager.DeletePrevious=true
@@ -261,23 +261,23 @@ RCC.CortexFreq_Value=64000000
 RCC.CpuClockFreq_Value=64000000
 RCC.D1CPREFreq_Value=64000000
 RCC.D2PPRE1=RCC_APB1_DIV2
-RCC.DFSDMACLkFreq_Value=49500000
+RCC.DFSDMACLkFreq_Value=48000000
 RCC.DFSDMFreq_Value=64000000
 RCC.DIVM1=1
 RCC.DIVM2=1
 RCC.DIVN1=24
 RCC.DIVN2=19
-RCC.DIVP1Freq_Value=99000000
+RCC.DIVP1Freq_Value=96000000
 RCC.DIVP2Freq_Value=76000000
 RCC.DIVP3Freq_Value=16125000
 RCC.DIVQ1=4
-RCC.DIVQ1Freq_Value=49500000
+RCC.DIVQ1Freq_Value=48000000
 RCC.DIVQ2Freq_Value=76000000
 RCC.DIVQ3Freq_Value=16125000
-RCC.DIVR1Freq_Value=99000000
+RCC.DIVR1Freq_Value=96000000
 RCC.DIVR2Freq_Value=76000000
 RCC.DIVR3Freq_Value=16125000
-RCC.FDCANFreq_Value=49500000
+RCC.FDCANFreq_Value=48000000
 RCC.FMCFreq_Value=64000000
 RCC.FamilyName=M
 RCC.HCLK3ClockFreq_Value=64000000
@@ -297,13 +297,13 @@ RCC.MCO2PinFreq_Value=64000000
 RCC.QSPIFreq_Value=64000000
 RCC.RNGFreq_Value=48000000
 RCC.RTCFreq_Value=32000
-RCC.SAI1Freq_Value=49500000
-RCC.SAI23Freq_Value=49500000
-RCC.SAI4AFreq_Value=49500000
-RCC.SAI4BFreq_Value=49500000
-RCC.SDMMCFreq_Value=49500000
-RCC.SPDIFRXFreq_Value=49500000
-RCC.SPI123Freq_Value=49500000
+RCC.SAI1Freq_Value=48000000
+RCC.SAI23Freq_Value=48000000
+RCC.SAI4AFreq_Value=48000000
+RCC.SAI4BFreq_Value=48000000
+RCC.SDMMCFreq_Value=48000000
+RCC.SPDIFRXFreq_Value=48000000
+RCC.SPI123Freq_Value=48000000
 RCC.SPI45Freq_Value=64000000
 RCC.SPI6Freq_Value=64000000
 RCC.SWPMI1Freq_Value=32000000
@@ -313,8 +313,8 @@ RCC.Tim2OutputFreq_Value=64000000
 RCC.TraceFreq_Value=64000000
 RCC.USART16Freq_Value=64000000
 RCC.USART234578Freq_Value=32000000
-RCC.USBFreq_Value=49500000
-RCC.VCO1OutputFreq_Value=198000000
+RCC.USBFreq_Value=48000000
+RCC.VCO1OutputFreq_Value=192000000
 RCC.VCO2OutputFreq_Value=152000000
 RCC.VCO3OutputFreq_Value=32250000
 RCC.VCOInput1Freq_Value=8000000
